fix: fixed major performance issue caused by react's strict-mode
