+++ dev-environment/package.json
@@ -14,7 +14,7 @@
     "@types/react-dom": "18.2.7",
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.19",
-    "gleamy": "1.1.4",
+    "gleamy": "1.1.5",
     "next": "13.4.19",
     "react": "18.2.0",
     "react-dom": "18.2.0",
