+++ packages/gleamy/package.json
@@ -1,7 +1,7 @@
 {
   "name": "gleamy",
   "description": "Create a reactive shiny element in different materials",
-  "version": "1.1.4",
+  "version": "1.1.5",
   "license": "MIT",
   "private": false,
   "sideEffects": false,
