+++ packages/gleamy/src/lib/components/base-materials/ParticleMaterial.tsx
@@ -2,7 +2,6 @@ import { type FC, useEffect, useCallback, useContext, useRef } from 'react';
 import { setDPI } from '../../utils/setDPI';
 import { type TParticle } from '../../../types';
 import { GleamyContext } from '../../provider';
-import { drawMaterial } from '../../utils/drawMaterial';
 
 interface CreateParticles {
   depth: number;
@@ -251,11 +250,31 @@ export const ParticleMaterial: FC<TParticle> = ({
   ]);
 
   useEffect(() => {
-    if (rendering) {
-      const fps = gleamyProvider?.fps || 60;
-      drawMaterial({ drawFunction: render, fps });
+    const fps = gleamyProvider?.fps || 60;
+
+    let animationFrameId = 0;
+    const interval = 1000 / fps;
+    let now;
+    let then = Date.now();
+    let delta;
+
+    function draw() {
+      now = Date.now();
+      delta = now - then;
+
+      if (delta > interval) {
+        then = now - (delta % interval);
+        rendering && render();
+      }
+      animationFrameId = window.requestAnimationFrame(draw);
     }
-  }, [gleamyProvider, render, rendering]);
+
+    draw();
+
+    return () => {
+      window.cancelAnimationFrame(animationFrameId);
+    };
+  }, [props]);
 
   return (
     <canvas
